requirements.txt

pydub
AppOpener
pyautogui
duckduckgo-search
ollama
lancedb
beautifulsoup4
sentence-transformers
mss
pillow
fastapi
uvicorn
